Fix: 추가정보 입력 로직 수정
