+++ pages/index.tsx
@@ -1,8 +1,3 @@
-import { useEffect, useState } from 'react';
-import { requestUserRole } from 'apis/auth';
-import PortalModal from 'components/commons/PortalModal';
-import ProfileModal from 'components/commons/modals/ProfileModal';
-import WelcomeModal from 'components/commons/modals/WelcomeModal';
 import HomeBanner from 'components/pages/home/HomeBanner';
 import HomeCardSlider from 'components/pages/home/HomeCardSlider';
 import HomeEditorRecommend from 'components/pages/home/HomeEditorRecommend';
@@ -12,26 +7,6 @@ import HomeMobileMagazineSlider from 'components/pages/home/HomeMobileMagazineSl
 import HomeSliderWithPagination from 'components/pages/home/HomeSliderWithPagination';
 
 const Home = () => {
-  const [isSignUpModalOpen, setIsSignUpModalOpen] = useState(false);
-  const [signUpStatus, setSignUpStatus] = useState('welcome');
-
-  const onNextClick = () => {
-    setSignUpStatus(() => 'signUp');
-  };
-
-  useEffect(() => {
-    const reqUserRole = async () => {
-      const roleRes = await requestUserRole();
-
-      // roleRes = "GUEST" | "USER" | "SIGNED_OUT"
-      const isSignUpNeeded = roleRes === 'GUEST';
-
-      setIsSignUpModalOpen(isSignUpNeeded);
-    };
-
-    reqUserRole();
-  }, []);
-
   return (
     <div className='mb-76 mt-76 flex  flex-col items-center justify-center gap-76 md:mt-0 md:items-start'>
       <HomeSliderWithPagination mode='hero' />
@@ -55,13 +30,6 @@ const Home = () => {
         <HomeMagazineSlider />
         <HomeMobileMagazineSlider />
       </div>
-      <PortalModal openStatus={isSignUpModalOpen}>
-        {signUpStatus === 'welcome' ? (
-          <WelcomeModal handleNextClick={onNextClick} />
-        ) : (
-          <ProfileModal setIsModalOpen={setIsSignUpModalOpen} />
-        )}
-      </PortalModal>
     </div>
   );
 };
